+++ framework/log/migrations/m141106_185632_log_init.php
@@ -66,8 +66,7 @@ class m141106_185632_log_init extends Migration
 
             $tableOptions = null;
             if ($this->db->driverName === 'mysql') {
-                // https://stackoverflow.com/questions/766809/whats-the-difference-between-utf8-general-ci-and-utf8-unicode-ci
-                $tableOptions = 'CHARACTER SET utf8 COLLATE utf8_unicode_ci ENGINE=InnoDB';
+                $tableOptions = sprintf('CHARACTER SET %s ENGINE=InnoDB', $this->db->effectiveCharset);
             }
 
             $this->createTable($target->logTable, [