+++ framework/rbac/migrations/m140506_102106_rbac_init.php
@@ -57,8 +57,7 @@ class m140506_102106_rbac_init extends \yii\db\Migration
         $tableOptions = null;
 
         if ($this->db->driverName === 'mysql') {
-            // https://stackoverflow.com/questions/766809/whats-the-difference-between-utf8-general-ci-and-utf8-unicode-ci
-            $tableOptions = 'CHARACTER SET utf8 COLLATE utf8_unicode_ci ENGINE=InnoDB';
+            $tableOptions = sprintf('CHARACTER SET %s ENGINE=InnoDB', $this->db->effectiveCharset);
         }
 
         $this->createTable(