+++ framework/web/migrations/m160313_153426_session_init.php
@@ -24,8 +24,7 @@ class m160313_153426_session_init extends Migration
         $tableOptions = null;
 
         if ($this->db->driverName === 'mysql') {
-            // https://stackoverflow.com/questions/766809/whats-the-difference-between-utf8-general-ci-and-utf8-unicode-ci
-            $tableOptions = 'CHARACTER SET utf8 COLLATE utf8_unicode_ci ENGINE=InnoDB';
+            $tableOptions = sprintf('CHARACTER SET %s ENGINE=InnoDB', $this->db->effectiveCharset);
         }
 
         $this->createTable(